First version of MHS for Windows Phone
First commit of MyHitchhikingSpots for Windows Phone. This should be the
version that was published on Windows Phone store.
There's an Android project here too, but it should be ignore since this
project was dropped in order to start a new one using Java instead of
Xamarin, C# and other Windows technologies.